Refactor SpotlightFilters code
