+++ components/SpotlightFilters/index.ts
@@ -1,1 +1,2 @@
 export * from './SpotlightFilters';
+export * from './utils';
